Move some code in more relevant modules.
