+++ test/dofiles.py
@@ -5,7 +5,8 @@ from __future__ import print_function, division, absolute_import, unicode_litera
 
 
 from settings import lab_and_model_settings, simulation_settings, my_cvspath
-from dr2xml import pingFileForRealmsList, generate_file_defs
+from dr2xml import generate_file_defs
+from pingfiles_interface import pingFileForRealmsList
 from vars_selection import select_CMORvars_for_lab
 from settings_interface import initialize_dict
